EXtract dirty node cache to its own file
